policies

+++ app/Http/Controllers/ImageController.php
@@ -9,7 +9,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class ImageController extends Controller
-{    
+{   
 
     /**
      * Remove the specified resource from storage.
@@ -17,8 +17,11 @@ class ImageController extends Controller
      * @param  \App\models\Image  $image
      * @return \Illuminate\Http\Response
      */
-    public function destroy(image $image)
-    {      
+    public function destroy(Request $request, image $image)
+    {
+        // Policies
+        if($request->user()->cant('delete',$image))
+            abort('403','No Autorizado para ver eliminar esta imagen');      
 
         $endisk = Storage::disk('public')->delete($image->address);
